+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 * cli: Update authentication examples to remove password flag and make
   subcommend selection a bit clearer
   ([PR](https://github.com/hashicorp/boundary/pull/1835))
+* Data Warehouse: Add addresses on plugin based hosts to the database warehouse.
+  3 new dimension tables have been added including `wh_network_address_group`
+  (which is now referenced by `wh_host_dimension`),
+  `wh_network_address_dimension`, and `wh_network_address_group_membership`.
+  ([PR](https://github.com/hashicorp/boundary/pull/1855))
 
 ### Bug Fixes
 * targets: Specifying a plugin based host id when authorizing a session